+++ src/authors/authors.controller.ts
@@ -1,29 +1,30 @@
-import { Body, Controller, Delete, Get, Param, Patch, Post } from '@nestjs/common';
+import { Body, Controller, Get, Param, Patch, Post } from '@nestjs/common';
 import { AuthorsService } from './authors.service';
 import { CreateAuthorDto } from './dto/create-author.dto';
 import { UpdateAuthorDto } from './dto/update-author.dto';
+import { Author } from './interfaces/author.interface';
 
 @Controller('authors')
 export class AuthorsController {
   constructor(private readonly authorsService: AuthorsService) {}
 
   @Get()
-  async getAuthors() {
+  async getAuthors(): Promise<Author[]> {
     return this.authorsService.getAuthors();
   }
 
   @Post()
-  async createAuthor(@Body() createAuthorDto: CreateAuthorDto) {
+  async createAuthor(
+    @Body() createAuthorDto: CreateAuthorDto,
+  ): Promise<Author> {
     return this.authorsService.createAuthor(createAuthorDto);
   }
 
   @Patch('update/:id')
-  async updateAuthor(@Param('id') id: number, @Body() updateAuthorDto: UpdateAuthorDto) {
-    return this.authorsService.updateAuthor(id, updateAuthorDto)
-  }
-
-  @Delete('delete/:id')
-  async deleteAuthor(@Param('id') id: number) {
-    return this.authorsService.deleteAuthor(id)
+  async updateAuthor(
+    @Param('id') id: number,
+    @Body() updateAuthorDto: UpdateAuthorDto,
+  ): Promise<Author> {
+    return this.authorsService.updateAuthor(id, updateAuthorDto);
   }
 }