+++ src/books/books.module.ts
@@ -1,13 +1,10 @@
 import { Module } from '@nestjs/common';
-import { AuthorsModule } from 'src/authors/authors.module';
-import { AuthorsService } from 'src/authors/authors.service';
-import { BooksAuthorService } from 'src/books-author/books-author.service';
 import { BooksController } from './books.controller';
 import { BooksService } from './books.service';
 
 @Module({
   controllers: [BooksController],
-  providers: [BooksService, BooksAuthorService],
-  imports: [AuthorsModule]
+  providers: [BooksService],
+  exports: [BooksService]
 })
 export class BooksModule {}