+++ CHANGELOG.md
@@ -5,7 +5,8 @@ Changelog Format
 NOTE: all lowercase and no ending punctuation marks
 -->
 
-- [CY] - [#4](https://github.com/cyril-deguzman/studyuwu/issues/3) app(service): add or remove author to and from books
+- [CY] - [#5](https://github.com/cyril-deguzman/studyuwu/issues/5) app(all): refactor whole app
+- [CY] - [#4](https://github.com/cyril-deguzman/studyuwu/issues/4) app(service): add or remove author to and from books
 - [CY] - [#3](https://github.com/cyril-deguzman/studyuwu/issues/3) app(module): implement basic crud operations, validation, and exception filters for books module
 - [CY] - [#2](https://github.com/cyril-deguzman/studyuwu/issues/2) app(module): implement basic crud operations, validation, and exception filters for authors module
 - [CY] - [#1](https://github.com/cyril-deguzman/studyuwu/issues/1) app(module/controller/dto): created book and author modules, controllers, and dtos